requirements.txt
aniso8601==7.0.0
asgiref==3.4.1
Django==3.2.7
djangorestframework==3.12.4
django-cors-headers==3.8.0
graphene==2.1.9
graphene-django==2.14.0
graphql-core==2.3.2
graphql-relay==2.0.1
promise==2.3
psycopg2==2.9.1
pytz==2021.1
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
